Use locale of moment instance, instead of global moment
Relying on the locale of the global moment singleton can run into
issues if NPM behaves unpredictably. Since we already can have an
instance passed in as a prop, we can just use the locale from that,
and since we're wrapping it with a moment object even if it *isn't*
passed in as a moment object, this should fall back to the global
moment object, which was the previous behavior.